# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
...
    Product:
      type: object
      required: 
        - barcode
        - name
        - category
        - weight
        - sellPrice
        - stock
      properties: 
        barcode: 
          type: string
          example: "6430039222212"
        name: 
          type: string
          example: Kombucha bakteerin makuinen bakteerijuoma
        category: 
          '$ref': '#/components/schemas/Category'
        weight: 
          type: integer
          example: 370
        sellPrice: 
          type: number
          example: 3.10
        stock: 
          type: integer
          example: 10
    AdminProduct:
      allOf:
        - '$ref': '#/components/schemas/Product'
        - type: object
          required:
            - buyPrice
          properties:
            buyPrice:
              type: number
              example: 3.00
...
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
        description: 
          type: string
          example: Drinks, soda with bottle return
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/products: 
    get: 
      summary: Get a list of all products
      security:
        - normalAccessToken: []
      responses: 
        200:
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - products
                properties: 
                  products: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Product'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/products/{barcode}: 
    get:
      summary: Get details of a specific product
      security:
        - normalAccessToken: []
      parameters: 
        - name: barcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema:
                type: object
                required: 
                  - product
                properties: 
                  product:
                    '$ref': '#/components/schemas/Product'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/products: 
    get: 
      summary: Get a list of all products
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - products
                properties: 
                  products: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    post: 
      summary: Create a new product
      security:
        - adminAccessToken: []
      responses: 
        201: 
          description: New product created successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - barcode
                - name
                - categoryId
                - weight
                - buyPrice
                - sellPrice
                - stock
              properties: 
                barcode: 
                  type: string
                  example: '7314660003114'
                name: 
                  type: string
                  example: Muumilimsa
                categoryId: 
                  type: integer
                  example: 10
                weight: 
                  type: number
                  example: 350
                buyPrice: 
                  type: number
                  example: 2.10
                sellPrice: 
                  type: number
                  example: 2.20
                stock: 
                  type: integer
                  example: 8
  /api/v1/admin/products/{barcode}: 
    get: 
      summary: Get details of a specific product
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update the details of a specific product
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              properties: 
                name: 
                  type: string
                  example: Muumilimsa
                categoryId: 
                  type: integer
                  example: 10
                weight: 
                  type: number
                  example: 350
                buyPrice: 
                  type: number
                  example: 2.10
                sellPrice: 
                  type: number
                  example: 2.20
                stock: 
                  type: integer
                  example: 8
    delete: 
      summary: Delete a product
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedProduct
                properties: 
                  deletedProduct: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...